pyproject.toml
[project]
name = "mmevallab"
version = "0.1.0"
description = "Unified multimodal regression harness for MMMU, OmniDocBench, Video-MME"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "Apache-2.0"}
authors = [{name = "MMEvalLab Team"}]
keywords = ["multimodal", "evaluation", "benchmark", "vlm", "regression"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "click>=8.0",
    "rich>=13.0",
    "tqdm>=4.65",
]

[project.optional-dependencies]
eval = [
    "datasets>=2.14",
    "pillow>=10.0",
    "numpy>=1.24",
]
video = [
    "av>=10.0",
    "opencv-python>=4.8",
]
pdf = [
    "pymupdf>=1.23",
]
faiss = [
    "faiss-cpu>=1.7",
]
dev = [
    "pytest>=7.4",
    "pytest-cov>=4.1",
    "ruff>=0.1",
    "mypy>=1.5",
    "pre-commit>=3.4",
]
all = [
    "mmevallab[eval,video,pdf,faiss,dev]",
]

[project.scripts]
mmeval = "mmevallab.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["mmevallab"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_ignores = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"
